+++ java-desktop-util-favicon/src/main/java/module-info.java
@@ -1,0 +1,17 @@
+
+module nbbrd.desktop.favicon {
+
+    requires static lombok;
+    requires static org.checkerframework.checker.qual;
+    requires static nbbrd.design;
+    requires static nbbrd.service;
+
+    requires transitive java.desktop;
+
+    exports nbbrd.desktop.favicon;
+
+    uses nbbrd.desktop.favicon.FaviconSupplier;
+    provides nbbrd.desktop.favicon.FaviconSupplier with
+            internal.desktop.favicon.FaviconkitSupplier,
+            internal.desktop.favicon.GoogleSupplier;
+}