+++ java-desktop-util-favicon/src/main/java/nbbrd/desktop/favicon/URLConnectionFactory.java
@@ -1,0 +1,18 @@
+package nbbrd.desktop.favicon;
+
+import lombok.NonNull;
+
+import java.io.IOException;
+import java.net.Proxy;
+import java.net.URL;
+import java.net.URLConnection;
+
+@FunctionalInterface
+public interface URLConnectionFactory {
+
+    @NonNull URLConnection openConnection(@NonNull URL url) throws IOException;
+
+    static @NonNull URLConnectionFactory getDefault() {
+        return URL::openConnection;
+    }
+}